fix(logger): move collision tests to src/__tests__/ and deduplicate
Move the 4 base field collision protection tests from
packages/logger/__tests__/ into packages/logger/src/__tests__/
so they are covered by the package tsconfig include path.
Delete the duplicate outer test file and unused helpers.

Closes #728
